Library as a separate file.
