+++ openfermionpsi4/_psi4_conversion_functions.py
@@ -18,8 +18,6 @@
 from __future__ import absolute_import
 
 import numpy
-import psi4
-import pubchempy as pcp
 
 from openfermion.ops import InteractionOperator
 
@@ -257,66 +255,4 @@ def parse_psi4_ccsd_amplitudes(number_orbitals,
 
     return single_amplitudes, double_amplitudes
 
-def parse_sdf_geometry(geom, n_atoms):
-    opf_geom = []
-    vec = geom.split('\n')[2:2 + n_atoms]
-
-    for i in range(len(vec)):
-        x = vec[i].split()[:4]
-        tup = tuple([float(x) for x in x[1:]])
-        atom = x[0].lower().capitalize()
-        opf_geom.append((atom, tup))
-
-    return opf_geom
-
-
-def parse_psi4_geometry(geom, n_atoms):
-    opf_geom = []
-    vec = geom.split('\n')[2:2 + n_atoms]
-
-    for i in range(len(vec)):
-        x = vec[i].split()[:4]
-        tup = tuple([float(x) for x in x[:3]])
-        atom = x[3].lower().capitalize()
-        opf_geom.append((atom, tup))
-    return opf_geom
-
-
-def extract(name_):
-    """Function to create MolecularData geometry from the molecule's name.
-
-    Args:
-        name_: a string giving the molecule's name as required by the PubChem
-            database. This can be quite flexible, e.g.: 'water' or 'H2O' or
-            or 'dihyrogen oxide' for the same molecule.
-
-    Returns:
-        opf_geom: a list of tuples giving the coordinates of each atom with
-        distances in Angstrom. Example is:
-        [('O', (0.0, 0.0, -0.066779921147764)),
-        ('H', (0.0, -0.763469300299257, 0.529922904804988)),
-        ('H', (-0.0, 0.763469300299257, 0.529922904804988))]
-    """
-    try:
-        name = pcp.get_compounds(name_, 'name', record_type='3d')[0]
-        mlc = psi4.geometry("""
-pubchem:{}
-""".format(name_))
-        mlc = mlc.create_psi4_string_from_molecule()
-        n_atoms = len(name.atoms)
-        opf_geom = parse_sdf_geometry(mlc, n_atoms)
-        print('3D compound found.')
-
-    except IndexError:
-        try:
-            name = pcp.get_compounds(name_, 'name', record_type='2d')[0]
-            n_atoms = len(name.atoms)
-            mlc = pcp.get_sdf(name.cid)
-            opf_geom = parse_psi4_geometry(mlc, n_atoms)
-            print('2D compound found.')
-        except IndexError:
-            opf_geom = 0
-            print("Unable to find molecule in the PubChem database.")
-
-    return opf_geom
